Rename AsmScope* to Scope*
