+++ libyul/Scope.h
