+++ libyul/ScopeFiller.cpp
@@ -19,10 +19,10 @@
  * Module responsible for registering identifiers inside their scopes.
  */
 
-#include <libyul/AsmScopeFiller.h>
+#include <libyul/ScopeFiller.h>
 
 #include <libyul/AST.h>
-#include <libyul/AsmScope.h>
+#include <libyul/Scope.h>
 #include <libyul/AsmAnalysisInfo.h>
 #include <libyul/Exceptions.h>
